requirements.txt
httpx>=0.27
pymongo>=4.6
adlfs==2024.12.0
azure-core==1.32.0
azure-datalake-store==0.0.53
azure-identity==1.20.0
azure-storage-blob==12.24.1
beautifulsoup4==4.13.3
fsspec==2025.2.0
prefect==3.2.7
python-dotenv==1.0.1
requests==2.32.3